refactor(accounting): compute money and quantity in Decimal on the live path
Stage 2 of the money-representation plan. The live account replay and book
fill compute in Decimal for exact arithmetic; the backtest stays float.

The shared per-fill math (ADR 020) is made generic over a constrained
Number = TypeVar("Number", float, Decimal): buy/sell deltas, apply_buy,
apply_sell, and the order validators. The backtest hot loop (simulation,
metrics) calls them with float and is unchanged, so the replay matches
byte for byte; the live replay calls them with Decimal.

- compute_account_state takes a Decimal initial_cash and returns a Decimal
  AccountState; apply_book_fill_transition is Decimal and returns a Decimal
  BookFillTransition.
- AccountState and BookFillTransition fields become Decimal.
- Live consumers (submission, analysis, reporting, web summaries) convert
  Decimal to float at their edge, because storage is still float until
  Stage 3 wires the encoder.
- sizing.py stays float this stage: it only picks a whole int until Stage 4.
- normalize_trade_fields stays a float row parser; the replay converts.

The whole-units guard stays until Stage 4. Storage is unchanged.

+++ src/trading/models/books.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import json
 from collections.abc import Iterator, Mapping
 from dataclasses import dataclass
+from decimal import Decimal
 
 from common.coercion import (
     row_expect_float,
@@ -422,20 +423,20 @@ class LedgerEntryRecord:
 class BookFillTransition:
     symbol: str
     side: str
-    qty: float
-    fill_price: float
-    commission: float
-    requested_price: float | None
-    cash_delta: float
-    realized_pnl_delta: float
-    slippage_amount: float
-    ending_qty: float
-    ending_avg_cost: float
-    ending_cash: float
-    ending_realized_pnl: float
-    ending_market_value: float
-    ending_unrealized_pnl: float
-    ending_equity: float
+    qty: Decimal
+    fill_price: Decimal
+    commission: Decimal
+    requested_price: Decimal | None
+    cash_delta: Decimal
+    realized_pnl_delta: Decimal
+    slippage_amount: Decimal
+    ending_qty: Decimal
+    ending_avg_cost: Decimal
+    ending_cash: Decimal
+    ending_realized_pnl: Decimal
+    ending_market_value: Decimal
+    ending_unrealized_pnl: Decimal
+    ending_equity: Decimal
 
 
 # --- Risk ---
